ATMA-18: migrated to std filesystem

+++ atomina-test/RandomGeneratorTest.cpp
@@ -31,14 +31,14 @@ namespace Random {
 
 		TEST_METHOD(Random_positive_float) {
 			ATMA::RandomGenerator<float> rg;
-			int value = rg(0.0f, 10.0f);
+			float value = rg(0.0f, 10.0f);
 
 			Assert::IsTrue(value >= 0.0f && value <= 10.0f);
 		}
 
 		TEST_METHOD(Random_negative_float) {
 			ATMA::RandomGenerator<float> rg;
-			int value = rg(-10.0f, 0.0f);
+			float value = rg(-10.0f, 0.0f);
 
 			Assert::IsTrue(value <= 0.0f && value >= -10.0f);
 		}
@@ -46,7 +46,7 @@ namespace Random {
 
 		TEST_METHOD(Random_mixed_float) {
 			ATMA::RandomGenerator<float> rg;
-			int value = rg(-10.0f, 10.0f);
+			float value = rg(-10.0f, 10.0f);
 
 			Assert::IsTrue(value >= -10.0f && value <= 10.0f);
 		}
@@ -54,14 +54,14 @@ namespace Random {
 
 		TEST_METHOD(Random_positive_double) {
 			ATMA::RandomGenerator<double> rg;
-			int value = rg(0.0, 10.0);
+			double value = rg(0.0, 10.0);
 
 			Assert::IsTrue(value >= 0.0 && value <= 10.0);
 		}
 
 		TEST_METHOD(Random_negative_double) {
 			ATMA::RandomGenerator<double> rg;
-			int value = rg(-10.0, 0.0);
+			double value = rg(-10.0, 0.0);
 
 			Assert::IsTrue(value <= 0.0 && value >= -10.0);
 		}
@@ -69,7 +69,7 @@ namespace Random {
 
 		TEST_METHOD(Random_mixed_double) {
 			ATMA::RandomGenerator<double> rg;
-			int value = rg(-10.0, 10.0);
+			double value = rg(-10.0, 10.0);
 
 			Assert::IsTrue(value >= -10.0 && value <= 10.0);
 		}
